Add Go bindings
This adds Go bindings for the QFS C API.  It implements some of the
available functions, specifically those which have been used for
internal services within Quantcast.  There are definitely some missing,
but this library may definitely be useful to others even in it's current
form.

All the code has been previously reviewed and approved inside Quantcast.
I'm removing the Git history to reduce the change of exposing
confidential information, but this library was implemented by the
following people:
* Steve Day
* Eric Culp
* Noah Goldman

This is currently being used in production at Quantcast.